backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
PyYAML>=6.0
mem0ai>=1.0.0
qdrant-client>=1.3.0
